Revert "WIP: Add a conversion object and  element based on the yuv crate"
This reverts commit f95b0d5099fa1991dead94af02f5d6fbe1daeaee.
