+++ Cargo.toml
@@ -63,8 +63,6 @@ members = [
     "video/gtk4",
     "video/hsv",
     "video/png",
-    "video/videoconvert",
-    "video/videoconvert/videoconverter",
     "video/rav1e",
     "video/skia",
     "video/videofx",
@@ -122,8 +120,6 @@ default-members = [
     "video/gif",
     "video/hsv",
     "video/png",
-    "video/videoconvert",
-    "video/videoconvert/videoconverter",
     "video/rav1e",
     "video/skia",
 ]
