+++ meson.build
@@ -218,7 +218,6 @@ plugins = {
     'library': 'libgstrspng',
     'examples': ['pngenc'],
   },
-  'videoconvert': {'library': 'libgstrsvideoconvert'},
   'rav1e': {'library': 'libgstrav1e'},
   'videofx': {
     'library': 'libgstrsvideofx',
