+++ meson_options.txt
@@ -68,7 +68,6 @@ option('gif', type: 'feature', value: 'auto', description: 'Build gif plugin')
 option('gtk4', type: 'feature', value: 'auto', description: 'Build GTK4 plugin')
 option('hsv', type: 'feature', value: 'auto', description: 'Build hsv plugin')
 option('png', type: 'feature', value: 'auto', description: 'Build png plugin')
-option('videoconvert', type: 'feature', value: 'auto', description: 'Build rsvideoconvert plugin')
 option('rav1e', type: 'feature', value: 'auto', description: 'Build rav1e plugin')
 option('skia', type: 'feature', value: 'auto', description: 'Build skia plugin')
 option('videofx', type: 'feature', value: 'auto', description: 'Build videofx plugin')
