+++ utils/segmentclipper/build.rs
@@ -1,3 +1,4 @@
 fn main() {
     gst_plugin_version_helper::info()
 }
+
