--- utils/validate/tests/rgb_to_yuv_conversions.validatetest
@@ -1,36 +1,0 @@
-# RGB→YUV format conversions test
-# Tests RGB→YUV conversions for all supported formats using videotestsrc with SMPTE pattern
-# Converts from RGB to various YUV formats and verifies by converting back to RGB
-
-meta,
-    args = {
-        "videotestsrc pattern=smpte num-buffers=1 ! video/x-raw,width=320,height=240,format=RGB ! tee name=t \
-         t. ! queue ! pngenc ! filesink location=\"$(tmpdir)/rgb_to_yuv_reference.png\" \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=NV12  ! fakesink name=nv12_rgb enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=NV21  ! fakesink name=nv21_rgb enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=I420  ! fakesink name=i420_rgb enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=YV12  ! fakesink name=yv12_rgb enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=Y42B  ! fakesink name=y42b_rgb enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=Y444  ! fakesink name=y444_rgb enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=NV16  ! fakesink name=nv16_rgb enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=NV61  ! fakesink name=nv61_rgb enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=NV24  ! fakesink name=nv24_rgb enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=GRAY8 ! fakesink name=gray8_rgb enable-last-sample=true",
-    },
-    handles-states=true,
-    ignore-eos=true
-
-play
-wait, message-type=eos
-
-# Test that RGB round-trip conversions match the reference
-# Threshold is slightly higher to account for YUV subsampling losses
-foreach, format=<nv12_rgb, nv21_rgb, i420_rgb, yv12_rgb, y42b_rgb, y444_rgb, nv16_rgb, nv61_rgb, nv24_rgb>,
-    actions = {
-        [compare-last-frame, reference-file="$(tmpdir)/rgb_to_yuv_reference.png", sink-name="$(format)", threshold=0.15, metric=rms],
-    }
-
-# Test grayscale with higher threshold (loses color information)
-compare-last-frame, reference-file="$(tmpdir)/rgb_to_yuv_reference.png", sink-name=gray8_rgb, threshold=0.3, metric=rms
-
-stop
