--- utils/validate/tests/rsvideoconvert.validatetest
@@ -1,36 +1,0 @@
-# Test rsvideoconvert element for RGB↔YUV conversions
-# Tests various format conversions and verifies output using compare-last-frame
-
-meta,
-    args = {
-        "videotestsrc pattern=smpte num-buffers=1 ! video/x-raw,width=320,height=240,format=RGB ! tee name=t \
-         t. ! queue ! fakesink name=rgb_ref enable-last-sample=true \
-         t. ! queue ! pngenc ! filesink location=\"$(tmpdir)/rsvideoconvert_reference.png\" \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=I420 ! rsvideoconvert ! video/x-raw,format=RGB ! fakesink name=i420_roundtrip enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=NV12 ! rsvideoconvert ! video/x-raw,format=RGB ! fakesink name=nv12_roundtrip enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=NV21 ! rsvideoconvert ! video/x-raw,format=RGB ! fakesink name=nv21_roundtrip enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=YV12 ! rsvideoconvert ! video/x-raw,format=RGB ! fakesink name=yv12_roundtrip enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=Y42B ! rsvideoconvert ! video/x-raw,format=RGB ! fakesink name=y42b_roundtrip enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=Y444 ! rsvideoconvert ! video/x-raw,format=RGB ! fakesink name=y444_roundtrip enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=NV16 ! rsvideoconvert ! video/x-raw,format=RGB ! fakesink name=nv16_roundtrip enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=NV24 ! rsvideoconvert ! video/x-raw,format=RGB ! fakesink name=nv24_roundtrip enable-last-sample=true \
-         t. ! queue ! rsvideoconvert ! video/x-raw,format=GRAY8 ! rsvideoconvert ! video/x-raw,format=RGB ! fakesink name=gray8_roundtrip enable-last-sample=true",
-    },
-    handles-states=true,
-    ignore-eos=true
-
-play
-wait, message-type=eos
-
-# Test round-trip conversions RGB→YUV→RGB
-# Threshold accounts for YUV subsampling losses
-
-foreach, format=<i420_roundtrip, nv12_roundtrip, nv21_roundtrip, yv12_roundtrip, y42b_roundtrip, y444_roundtrip, nv16_roundtrip, nv24_roundtrip>,
-    actions = {
-        [compare-last-frame, reference-file="$(tmpdir)/rsvideoconvert_reference.png", sink-name="$(format)", threshold=0.15, metric=rms],
-    }
-
-# Test grayscale with higher threshold (loses color information)
-compare-last-frame, reference-file="$(tmpdir)/rsvideoconvert_reference.png", sink-name=gray8_roundtrip, threshold=0.3, metric=rms
-
-stop
