--- utils/validate/tests/rsvideoconvert_simple.validatetest
@@ -1,13 +1,0 @@
-# Simple test for rsvideoconvert element
-# Tests basic RGB→I420 conversion
-
-meta,
-    args = {
-        "videotestsrc pattern=smpte num-buffers=1 ! video/x-raw,width=320,height=240,format=RGB ! rsvideoconvert ! video/x-raw,format=I420 ! fakesink name=sink enable-last-sample=true",
-    },
-    handles-states=true,
-    ignore-eos=true
-
-play
-wait, message-type=eos
-stop
