--- video/videoconvert/Cargo.toml
@@ -1,45 +1,0 @@
-[package]
-name = "gst-plugin-videoconvert"
-version.workspace = true
-authors = ["Thibault Saunier <tsaunier@igalia.com>"]
-repository.workspace = true
-license = "MPL-2.0"
-edition.workspace = true
-rust-version.workspace = true
-description = "GStreamer Rust Video Format Converter"
-
-[dependencies]
-gst.workspace = true
-gst-base.workspace = true
-gst-video.workspace = true
-gst-videoconverter = { path = "./videoconverter" }
-
-[dev-dependencies]
-gst-check.workspace = true
-
-[lib]
-name = "gstrsvideoconvert"
-crate-type = ["cdylib", "rlib"]
-path = "src/lib.rs"
-
-[build-dependencies]
-gst-plugin-version-helper.workspace = true
-
-[features]
-static = []
-capi = []
-doc = ["gst/v1_18"]
-
-[package.metadata.capi]
-min_version = "0.9.21"
-
-[package.metadata.capi.header]
-enabled = false
-
-[package.metadata.capi.library]
-install_subdir = "gstreamer-1.0"
-versioning = false
-import_library = false
-
-[package.metadata.capi.pkg_config]
-requires_private = "gstreamer-1.0, gstreamer-base-1.0, gstreamer-video-1.0, gobject-2.0, glib-2.0, gmodule-2.0"
