--- video/videoconvert/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    gst_plugin_version_helper::info()
-}
