--- video/videoconvert/src/rsvideoconvert/mod.rs
@@ -1,19 +1,0 @@
-// SPDX-License-Identifier: MPL-2.0
-
-use gst::glib;
-use gst::prelude::*;
-
-mod imp;
-
-glib::wrapper! {
-    pub struct RsVideoConvert(ObjectSubclass<imp::RsVideoConvert>) @extends gst_video::VideoFilter, gst_base::BaseTransform, gst::Element, gst::Object;
-}
-
-pub fn register(plugin: &gst::Plugin) -> Result<(), glib::BoolError> {
-    gst::Element::register(
-        Some(plugin),
-        "rsvideoconvert",
-        gst::Rank::NONE,
-        RsVideoConvert::static_type(),
-    )
-}
