--- video/videoconvert/videoconverter/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "gst-videoconverter"
-version.workspace = true
-authors = ["Thibault Saunier <tsaunier@igalia.com>"]
-repository.workspace = true
-license = "MPL-2.0"
-edition.workspace = true
-rust-version.workspace = true
-description = "GStreamer Rust Video Format Converter Library"
-
-[dependencies]
-gst.workspace = true
-gst-base.workspace = true
-gst-video.workspace = true
-yuv = {version = "0.8", features = ["rayon"]}
